+++ src/main/java/com/moing/backend/domain/board/application/service/UpdateBoardUseCase.java
@@ -4,7 +4,7 @@ import com.moing.backend.domain.board.application.dto.request.UpdateBoardRequest
 import com.moing.backend.domain.board.application.dto.response.UpdateBoardResponse;
 import com.moing.backend.domain.board.exception.NotAuthByBoardException;
 import com.moing.backend.global.response.BaseBoardServiceResponse;
-import com.moing.backend.global.util.BaseBoardService;
+import com.moing.backend.global.utils.BaseBoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
