+++ src/main/java/com/moing/backend/domain/boardComment/application/service/DeleteBoardCommentUseCase.java
@@ -5,7 +5,7 @@ import com.moing.backend.domain.boardComment.domain.service.BoardCommentDeleteSe
 import com.moing.backend.domain.boardComment.domain.service.BoardCommentGetService;
 import com.moing.backend.domain.boardComment.exception.NotAuthByBoardCommentException;
 import com.moing.backend.global.response.BaseBoardServiceResponse;
-import com.moing.backend.global.util.BaseBoardService;
+import com.moing.backend.global.utils.BaseBoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
