+++ src/main/java/com/moing/backend/domain/boardComment/application/service/GetBoardCommentUseCase.java
@@ -3,7 +3,7 @@ package com.moing.backend.domain.boardComment.application.service;
 import com.moing.backend.domain.boardComment.application.dto.response.GetBoardCommentResponse;
 import com.moing.backend.domain.boardComment.domain.service.BoardCommentGetService;
 import com.moing.backend.global.response.BaseBoardServiceResponse;
-import com.moing.backend.global.util.BaseBoardService;
+import com.moing.backend.global.utils.BaseBoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
