+++ src/main/java/com/moing/backend/domain/boardComment/domain/entity/BoardComment.java
@@ -3,7 +3,7 @@ package com.moing.backend.domain.boardComment.domain.entity;
 import com.moing.backend.domain.board.domain.entity.Board;
 import com.moing.backend.domain.teamMember.domain.entity.TeamMember;
 import com.moing.backend.global.entity.BaseTimeEntity;
-import com.moing.backend.global.util.AesConverter;
+import com.moing.backend.global.utils.AesConverter;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
